+++ barapost/barapost-local.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "3.15.c"
+__version__ = "3.16.a"
 # Year, month, day
 __last_update_date__ = "2020-09-21"